chore: restructure folder / files in 360 image package
