+++ viewer/packages/360-images/index.ts
@@ -2,11 +2,11 @@
  * Copyright 2022 Cognite AS
  */
 
-export { Image360 } from './src/Image360';
-export { Image360Collection } from './src/Image360Collection';
+export { Image360 } from './src/entity/Image360';
+export { Image360Collection } from './src/collection/Image360Collection';
 export { Image360Facade } from './src/Image360Facade';
-export { Image360CollectionFactory } from './src/Image360CollectionFactory';
-export { Image360Entity } from './src/Image360Entity';
-export { Image360Visualization } from './src/Image360Visualization';
+export { Image360CollectionFactory } from './src/collection/Image360CollectionFactory';
+export { Image360Entity } from './src/entity/Image360Entity';
+export { Image360Visualization } from './src/entity/Image360Visualization';
 export { Image360EnteredDelegate, Image360ExitedDelegate } from './src/types';
-export { DefaultImage360Collection } from './src/DefaultImage360Collection';
+export { DefaultImage360Collection } from './src/collection/DefaultImage360Collection';
