+++ viewer/packages/360-images/src/collection/Image360CollectionFactory.ts
@@ -6,9 +6,9 @@ import { Image360Descriptor, Image360Provider } from '@reveal/data-providers';
 import { SceneHandler } from '@reveal/utilities';
 import zip from 'lodash/zip';
 import { DefaultImage360Collection } from './DefaultImage360Collection';
-import { Image360Entity } from './Image360Entity';
-import { Image360Icon } from './Image360Icon';
-import { Image360CollectionIcons } from './visuals/Image360CollectionIcons';
+import { Image360Entity } from '../entity/Image360Entity';
+import { Image360Icon } from '../entity/Image360Icon';
+import { Image360CollectionIcons } from './Image360CollectionIcons';
 
 export class Image360CollectionFactory<T> {
   private readonly _image360DataProvider: Image360Provider<T>;
