+++ viewer/packages/360-images/src/collection/image360Icon.vert
