+++ viewer/packages/360-images/src/entity/Image360.ts
