+++ viewer/packages/360-images/src/entity/Image360Entity.ts
