+++ viewer/packages/api/src/public/migration/Cognite3DViewer.ts
@@ -64,10 +64,9 @@ import {
   determineResolutionCap,
   determineSsaoRenderParameters
 } from './renderOptionsHelpers';
-import { Image360Collection, Image360Entity } from '@reveal/360-images';
+import { Image360Collection, Image360Entity, Image360 } from '@reveal/360-images';
 import { Image360ApiHelper } from '../../api-helpers/Image360ApiHelper';
 import html2canvas from 'html2canvas';
-import { Image360 } from '@reveal/360-images/src/Image360';
 
 type Cognite3DViewerEvents = 'click' | 'hover' | 'cameraChange' | 'beforeSceneRendered' | 'sceneRendered' | 'disposed';
